Opt-out targets from using go/google3-java-launcher that are used at build time or depended on by the launcher itself.
When we flipped the launcher default to go/google3-java-launcher, we accidentally didn't include java_binary rules that are used as "tools" in genrules. Now we are trying to do this (see the bug referenced by this CL). When we enable this, there are various dependency loops caused by certain java_binary targets depending on go/google3-java-launcher.

This CL fixes one part of that dependency loop so that we can enable go/google3-java-launcher for all of the other genrule targets in google3. It should be entirely safe--if your target is only used at build time (which is true of almost all the targets being opted out in this set of CLs) then this CL is a no-op--it actually keeps the existing behavior of your target.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=122101949