Adding new method to query new breakpoint updates.
Currently, we use dequeueBreakpointUpdates() as both detection and extraction of updates. As we consider moving breakpoint updates serialization out of request lifetime (for better performance of the request), we need a query function that detects if there are any breakpoint updates during the request so we can carry out this information (and use it later to extract those updates).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=135255108