Look in more locations for source context.
Currently, the code only searches for source contexts in the same places it looks for java classes. The source context may be deployed to the app.yaml directory for a java application as part of the deployment process, and a customer may explicitly generate a source context next to the jar. This change searches the jar location and its parent app.yaml directory for the source context.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=135866712