Ignore source line information coming from classes with the ACC_SYNTHETIC flag.
When a class is marked with the ACC_SYHTHETIC flag, it means the class is a compiler-generated class that does not have actual source code. In this case, the source line table we get from the class file (which contains an entry that indicates which line the class was generated from) should be discarded, so that breakpoints should not be accidentally mapped to those synthetic classes.

As a specific example, the javac compiler generates inner classes for 'switch' statements, and the first line of a switch statement maps to both:
  (1) the outer class that has the actual switch statement,
  (2) the inner synthetic class that has some static initialization at that line.
Clearly, we don't want to map a breakpoint to (2), because that is not what the
user intended.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=153724359