Better format Class.Method in stack trace for Scala singleton/companion objects.
Scala has a singleton/companion object concept. All static methods and fields are placed in this singleton/companion object. It's syntax is as follows:

  object HelloWorld {
    ...
  }

In this case, the Scala compiler generates
   HelloWorld.class   => For the instance objects.
   HelloWorld$.class  => For the singleton/companion object.

The signature of a singleton object looks like this:
    "Lcom/prod/HelloWorld$;"

And we want to display its methods as "HelloWorld$.foo" rather than "HelloWorld..foo" in the stack trace. So, we should not be replacing the '$' with '.' in this case.

Implementation note: Nested objects become "LOuterClass$InnerObject$;". Hence, the '$' at the end an object signature is always followed by a ';'.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=154220787