Add basic protection against sun.misc.Unsafe by blocking breakpoints set inside classes that invoke any methods from sun.misc.Unsafe and com.google.protobuf.UnsafeUtil.
We perform a Java bytecode analysis at breakpoint set time (not when the class
is loaded, or while the application threads are stopped). If the bytecode analysis
finds a method call into the blacklisted classes, we mark the class as
unsafe, and fail the breakpoint with a BREAKPOINT_SOURCE_LOCATION error.

Note that this is not a bulletproof mechanism, yet it is expected to protect against most common uses of sun.misc.Unsafe.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158771952