[Java GCP] Remove data already passed in the debuggee from uniquifier computation.
This change is incompatible with previous versions of the agent as they would generate different debuggee-id for the same debugge. The major version for the agent should be bumped in the next release.

Note that the server hashes the entire debuggee proto to generate the debuggee-id.  Thus, hashing fields already passed in the debuggee into the uniquifier is redundant. This cl simplifies the code a bit for that purpose.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=163023379