Upgrade the ASM dependency of the Java Debugger open source build from version 5.0.4 to 6.0 to be able to handle classes built with JDK9.
The asm-6.0 jar we depend on contains a module-info.class that was built with class file format version 53 (the rest of the .class files in the jar are built with version 49). The maven-enforcer-plugin does not accept this .class file (we were using this plugin to make sure we don't accidentally pull any classes built with above Java7). Therefore, this CL also removes the use of the maven-enforcer-plugin until we can find another solution.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=174508280