Explicitly depend on javax.xml.bind in Java Debugger open source build. This is not about using the debugger with JDK9, but it's about building the debugger from open source using JDK9.
This is a no-op when building the debugger agent using JDK7-8. It is required with JDK9, because this dependency was moved into an external module in JDK9.

This has the diffbase, because jaxb-api also comes with a module-info.class file, which is incompatible with maven-enforcer-plugin.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=174512537