Handle Signature Polymorphic Java methods in Safe Method Caller.
+++ Details +++

When the Safe Caller encounters an "invoke" instruction, it looks for the target class and method signature in the constant pool. Then, it loads the class (into safe caller) and then searches for this method in that class.

With "Signature Polymorphic" methods, the method signature in the constant pool does not match the method signature inside the loaded class. That's why, in this CL, we also check if an equivalent method with a "Polymorphic Signature" exists.

+++ What's a "Signature Polymorphic" Method? +++

(from Java Virtual Machine Specification - Java SE 9 Edition)

A method is signature polymorphic if all of the following are true:
  1. It is declared in the java.lang.invoke.MethodHandle class or the java.lang.invoke.VarHandle class.
  2. It has a single formal parameter of type Object[].
  3. It has the ACC_VARARGS and ACC_NATIVE flags set.

The Java Virtual Machine gives special treatment to signature polymorphic methods in the invokevirtual instruction, in order to effect invocation of a method handle or to effect access to a variable referenced by an instance of java.lang.invoke.VarHandle.

+++ Example +++

The AtomicBoolean.compareAndGet(boolean, boolean) in JDK9 contains the following instruction:

    invokevirtual #4

Constant Pool contains the following:

     #4 // Method VarHandle.compareAndSet:(LAtomicBoolean;II)Z

However, the actual signature of compareAndSet method in the VarHandle class is:

  public native boolean java/lang/invoke/VarHandle.compareAndSet(java.lang.Object...);

Which is actually "Signature Polymorphic".

In this CL, we:
  1. First search 'VarHandle.compareAndSet' with method signature (LAtomicBoolean;II)Z.
     => Not found.
  2. Then search 'VarHandle.compareAndSet' with method signature ([LObject;)
     => Found.

+++ Impact on End User +++

Before, expressions were failing with this error:

   (Internal error executing AtomicBoolean.getAndSet at nanojava_interpreter.cc:205:
    failed to read instruction at offset 13, call stack:  ...)

Now, the same expression fails with:

   (Native method call to java.lang.invoke.VarHandle.compareAndSet is not allowed)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=177360262