Add an explicit dependency on //third_party/java/jdk:jaxb to code using javax.xml.bind
//third_party/java/jdk:jaxb is a neverlink target that provides the exact same
version of these classes as are found in the Google JDK 8, so this change is a
no-op.

It is being performed in preparation for JDK 9, which no longer provides these
classes. It will be necessary to migrate to a separately packaged,
non-neverlink, version of jaxb in //third_party/java/jaxb. This code is not yet
compatible with the separately packaged jaxb version, and will needed to be
fixed in preparation for Java 9.

Tested:
    tap_presubmit: http://test/OCL:190101029:BASE:190088286:1521768395317:a036d9c2
    Some tests failed; test failures are believed to be unrelated to this CL

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=190258686