C++: Clean up direct labels in rules
Instead of assuming which linker options are a label we use $(location label).
Figuring out what is a label is deprecated and will be disabled in a future
Blaze release.

This is a safe change (this can be proved by running blaze query before and after the change).

go/removing-linkopts-implicit-labels

Tested:
    TAP --sample for global presubmit queue
    http://test/OCL:191431935:BASE:192252573:1523358387608:62b6a61e

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=192316699