LSC: Relocate ABSL_FLAG statements outside of namespaces.
Placing Abseil Flags inside of namespace is not supported (it was never supported in fact) and leads to undefined behavior. This CL does not make any semantic changes. It only moves flags definitions outside of first top level namespace. If you observe the link failure caused by this change, it is an indication of duplicate flag definition.

More information is here: go/fix_namespaced_flag

Tested:
    tap_presubmit: http://test/OCL:237117311:BASE:238468016:1552614612270:13ca222d
    Some tests failed; test failures are believed to be unrelated to this CL
PiperOrigin-RevId: 238673108
Change-Id: I91c981ad6f230b92744a9d8081e42edb5bcef689