feat: Add active debuggee support.
The Firebase client now adds two new fields to the Debuggee entry in
the Firebase RTDB, `registrationTimeMsec` and `lastUpdateTimeMsec`.
The first time the Debuggee entry is written to the RTDB, both fields
are set. Thereafter only the `lastUpdateTimeMsec` will be changed, which
is done periodically every hour when the agent is running continuously,
it is also updated when the agent first starts up.
This is in support of
https://github.com/GoogleCloudPlatform/snapshot-debugger/issues/76.