feat: Add slf4j-simple dependency to internals.jar
The Firebase Admin SDK uses slf4j for logging. What this means is in
order to see logging information out of the sdk library a sl4j bind must
be found.

Here is we use the slf4j-simple logger, which emits the logs to stderr.
One simply way to configure it is to add
`-Dorg.slf4j.simpleLogger.defaultLogLevel=debug` to the java command
invocation.

See here for more information:
* https://www.slf4j.org/api/org/slf4j/simple/SimpleLogger.html
* https://www.slf4j.org/manual.html

To note, this is bundled into the agent's internals.jar file that the
agent loads for itself and is kept isolated from the application.